chore: rename proxy.ts to create-t.ts
