+++ src/index.ts
@@ -1,8 +1,8 @@
 export { Canvas, type CanvasProps } from "./canvas.tsx"
 export { Portal, Primitive } from "./components.tsx"
 export { $S3C } from "./constants.ts"
+export { createT } from "./create-t.tsx"
 export { useFrame, useLoader, useThree } from "./hooks.ts"
 export { applyProps } from "./props.ts"
-export { createT } from "./proxy.tsx"
 export * as S3 from "./types.ts"
 export { buildGraph } from "./utils/build-graph.ts"
